requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.85.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.4
inflection==0.5.1
kombu==5.2.4
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.10.2
PyJWT==1.7.1
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
